api-adaptation/requirements.txt
requests
loguru~=0.7.2
retry~=0.9.2
python-dotenv~=1.0.0
fastapi~=0.105.0
uvicorn~=0.24.0.post1
cos-python-sdk-v5
pydantic~=2.5.2
starlette~=0.27.0
sqlalchemy~=2.0.22
jsonpath-ng
pytz
gradio
db_hammer